move overlay selection top app bar to own file
